refactor(gpu-backends): extract Vulkan memory allocation and mapping into memory module
Rollback trigger: Kernel panic, stall > 1ms, test failure
